fix(image): use string types to load frames, fix warning in texture extraction

+++ cpu-sources/includes/texture/featuresExtractionT.h
@@ -5,4 +5,4 @@
 #include "image.h"
 #include "grayscale.h"
 
-uint8_t getVector(shared_image image, int x, int y);
+uint8_t getVector(shared_image image, size_t x, size_t y);
